backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
sentence-transformers
numpy
python-dotenv
pandas
